Sunspec: allow model fallback
