+++ util/modbus/modbus_test.go
@@ -1,30 +1,26 @@
 package modbus
 
-import "testing"
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
 
 func TestParsePoint(t *testing.T) {
 	tc := []struct {
-		in           string
-		model, block int
-		point        string
-		err          bool
+		in  string
+		ops []SunSpecOperation
 	}{
-		{"103:W", 103, 0, "W", false},
-		{"802:1:V", 802, 1, "V", false},
-		{"802::V", 802, 1, "V", true},
+		{"103:W", []SunSpecOperation{{103, 0, "W"}}},
+		{"802:1:V", []SunSpecOperation{{802, 1, "V"}}},
+		{"101|103:DCW", []SunSpecOperation{{101, 0, "DCW"}, {103, 0, "DCW"}}},
 	}
 
 	for _, tc := range tc {
 		t.Log(tc)
 
-		model, block, point, err := ParsePoint(tc.in)
-
-		if (err != nil) != tc.err {
-			t.Errorf("unexpected error: %d:%d:%s %v", model, block, point, err)
-		}
-
-		if !tc.err && (model != tc.model || block != tc.block || point != tc.point) {
-			t.Errorf("unexpected result: %d:%d:%s", model, block, point)
-		}
+		ops, err := ParsePoint(tc.in)
+		require.NoError(t, err)
+		require.Equal(t, tc.ops, ops)
 	}
 }
